+++ forte/fci/string_vo_list.cc
@@ -97,7 +97,7 @@ void StringLists::make_vo(std::shared_ptr<StringAddress> addresser, VOList& list
     auto b_begin = b.begin();
     auto b_end = b.begin() + n;
     if ((k >= 0) and (k <= n)) { // check that (n > 0) makes sense.
-        for (int h = 0; h < nirrep_; ++h) {
+        for (size_t h = 0; h < nirrep_; ++h) {
             // Create the key to the map
             std::tuple<size_t, size_t, int> pq_pair(p, q, h);
